CI: Merge configuration
Update the CI definition to contain all needed runs (baseline, private,
MR) to save on files that need to be maintained.

The updated file uses templates to make it easier to understand the
differences between all runs and hide the common configuration people
don't need to bother with. It also moves the realtime pipelines to a
separate branch so it's easier to distinguish it.

Bugzilla: INTERNAL
Upstream Status: RHEL-only

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>