memcg, slab: Fix incorrect placement of rcu_head in struct memcg_cache_params
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1951810
Upstream: RHEL only

The RHEL commit c65e7cba1b2e ("[mm] memcg, slab: RCU protect memcg_params
for root caches") incorrectly put rcu_head below memcg_caches[0] in
the memcg_cache_params structure. This makes rcu_head overlaps with the
memcg_caches[] array. Since the size of rcu_head is 2 long integers. It
overlaps with the first two entries of memcg_caches[].

When kfree_rcu() is called to free the memcg_cache_params structure,
the first two entries of memcg_caches[] are immediately overwritten
defeating the purpose of RCU protection. Fix that by moving rcu_head
above memcg_caches[0] as in the upstream patch.

A test kernel including this fix was delivered to a customer for
testing. As noted in comment #21 of the BZ, it was confirmed that the
test kernel did not crash under the scenario that previously crashed 100%
with normal kernel. This confirms that the fix is working.

Signed-off-by: Waiman Long <longman@redhat.com>