kernel/timer: Fix incorrect assertion in requeue_timers()
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2048502
Upstream Status: RHEL only

The assertion that the timer_jiffies in tvec_base should be equal to the
earliest expiry time of all the timers there does not really hold. So
remove that assertion.

In addition, the earliest expiry time could still be quite smaller
than the current jiffies in the case of a deferrable timer leading to
significant time in iterating the timer loop. So a lock-up like that
described in BZ #1849716 may still be possible. Force the timer_jiffies
to the current jiffies and let the expired timers be queued to the
current timer_jiffies slot (see __internal_add_timer()).

Fixes: 3e225ae62d9b ("[kernel] timer: Fix potential bug in requeue_timers()")
Signed-off-by: Waiman Long <longman@redhat.com>