mm: memcg: do not fail __GFP_NOFAIL charges
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2054345
Upstream Status: RHEL only

mem_cgroup_do_charge() needs to allow for overcharging the cgroup
in case the allocation is considered critical and the requestor
cannot cope with its failure (__GFP_NOFAIL). We take a simpler
approach in this rhel-only patch that is based on upstream commit
0d53c748bc9 ("memcg: ratify and consolidate over-charge handling"),
in order to reduce the code churn of a much bigger (and intrusive)
changeset.

Signed-off-by: Rafael Aquini <aquini@redhat.com>