xfs: use length to balance duplicate bno buffers in perag rb_tree
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2050464
Upstream Status: RHEL only
Tested: via fstests for regression and custom reproducer to
	instrument problematic state.

The XFS buffer cache is implemented as an rbtree indexed by disk
block address (i.e. ->b_bn). The combination of stale and variably
sized buffers introduces the transient possibility of buffers with
duplicate indexes. For example, a preexisting stale buffer that
happens to match the bno and length of a lookup request can be
reused, but if the buffer size of the lookup request differs then a
new buffer can be inserted to the tree with a duplicate index. The
location of this new buffer is hardcoded to pass to the right branch
of the existing set of buffers with a duplicate ->b_bn index.

This is historically not a problem in most cases because stale
buffers tend to be reclaimed aggressively once no longer referenced
and subsequent lookups use the same hardcoded algorithm to locate
the newly allocated buffer. This does not accommodate the fact that
the rbtree can rebalance in a way such that the hardcoded algorithm
fails, however. For example, if two buffers with matching ->b_bn
values but different lengths are inserted and staled in succession,
we can expect to see buffer B at the right branch of parent buffer
A. If the tree is rebalanced, it is possible to rotate B up a level
such that A rotates down to the left child of B.

At this stage, a lookup of buffer A will never succeed because the
lookup algorithm lands on B and expects any duplicate index buffers
to exist down the right side of the tree. If a lookup occurs as part
of a new allocation that happens to match the bno/length of buffer
A, this results in insertion of duplicate buffer A' in a separate
sub-branch of the tree from buffer A. This is essentially a cache
corruption at this point, but still may not yet result in observable
problems because subsequent lookups of the recently inserted buffer
follow the same path used for insertion (i.e. non-stale buffer A' is
reliably accessible). However if another rotation occurs that pushes
stale buffer A back up a level (so B is now a right child of A),
this creates a scenario where stale buffer A is an exact duplicate
and in the lookup path of recently added non-stale buffer A'.
Results are unpredictable from this point forward as lookups start
to return incorrect buffers.

This duplicate stale/non-stale buffer problem has been observed in
post-crash dump analysis of filesystems that have exhibited
transient corruption errors that otherwise do not reflect corruption
on-disk. The path to this condition described above has been
demonstrated in test via a combination of deliberate filesystem
changes and instrumented stale buffer removals that trigger
predictable rbtree rotations and subsequently result in spurious
buffer read verifier errors.

To resolve this problem, update the buffer lookup algorithm to
locate buffers with duplicate disk addresses by using buffer length
as a sub-index. This guarantees that a matching buffer bno/length
pair is always indexable and thus prevents insertion of duplicate
buffers into the tree.

Signed-off-by: Brian Foster <bfoster@redhat.com>