x86/speculation: cope with spectre_v2=retpoline cmdline on retbleed-affected Intel CPUs
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2090227
CVE: CVE-2022-23816
CVE: CVE-2022-23825
CVE: CVE-2022-29900
CVE: CVE-2022-29901

Upstream status: RHEL only

From Joe Mario (over e-mail):
"
Back in 2018 when RHEL was the only distro that defaulted to IBRS on
Skylake, we got lots of complaints from customers on why we were slower
than all the other distros. I raised the issue with the RHEL team and got
the OK to back out IBRS. They they decided to modify the RHEL-7 installer
so that it would add 'spectre_v2=retpoline' to new installations of
RHEL-7.7 and beyond.
"

With the advent of RETBleed IBRS becomes the defacto speculation control
protection mechanism for Intel CPUs affected by it and the decision we
made to back out IBRS comes back to haunt us (and our customers) as
follows:

1) all new RHEL-7 installations on Skylake CPUs, from 7.7 forward, are
   effectively disabling IBRS due to the added 'spectre_v2=retpoline'
   cmdline selection, leaving the system vulnerable to retbleed, and
   given that there's no effective way to revert the change to Anaconda
   because the GA images are not recreated, any new RHEL7 deploy on a
   Skylake CPU will also be vulnerable right out of the box;

2) any Skylake system with 'spectre_v2=retpoline' on its kernel cmdline
   will remain vulnerable, even after updating the kernel to a
   retbleed-safe build, without a suitable work-around;

In order to address this issue transparently, we're introducing this patch
that overrides 'spectre_v2=retpoline' on any Intel CPU flagged as affected
by RETBleed, making it behave as we originally introduced the speculative
control protection code and pick IBRS automatically. This way customers
that want to keep the mitigations ON and maintain their systems protected
can do so by just updating to the latest kernel build we release after this
particular change set.

Because the back-out decision was based, back in the day, on performance
concerns and complaints, we're offering an alternative way to customers
that want some protection against Spectre v2 but still think IBRS is too
taxing in performance, which is the newly 'spectre_v2=retpoline,force'
added cmdline option. Although we recommend against utilizing it, as
the system will remain vulnerable to RETBleed, we understand that some
customers might want to keep the old behavior at all costs.

Special thanks to Joe Mario, Josh Poimboeuf and Waiman Long, for
bringing up the problem and helping on refining the solution we
are presenting here.

Signed-off-by: Rafael Aquini <aquini@redhat.com>