mm: swap: disable swap_vma_readahead for PPC64
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2142455
Upstream Status: RHEL only

VMA-based swap readahead changes the default swap readahead algorithm to
read in blocks from the swap device following the memory access sequence
as opposed to consecutively read in blocks from the swap device which only
reflects the page-out order. Unfortunately, the feature can trip on itself
when running on PPC64 hosts because of RHEL7 support for non-linear mappings,
and the ambiguity that follows-up from that fact due to the overloading of
_PAGE_USER page-tables bit for that purpose in the PPC arch:

@ powerpc/include/asm/pte-hash64.h:
...
 17 #define _PAGE_PRESENT           0x0001 /* software: pte contains a translation */
 18 #define _PAGE_USER              0x0002 /* matches one of the PP bits */
 19 #define _PAGE_FILE              0x0002 /* (!present only) software: pte holds file offset */
...

because VMA-based RA code path do_swap_page_readahead()->swap_ra_info()
re-evaluates the PTE, it is susceptible to swap-in races which can lead to
the assertion at pte_to_swp_entry() to trigger as it checks for _PAGE_FILE on
the (now) swapped-in PTE and the overloading with _PAGE_USER will falsely look
like an inconsistency in the page-tables only on PowerPC systems.

Getting rid of the assertion without back-porting the removal for non-linear
mappings support is not the best option, nor it is a good option to force
check for (!_PAGE_PRESENT && _PAGE_FILE) in pte_file() -- as call sites of the
latter are not prepared to cope with it returning NULL. Since this is a known
and documented issue on PPC64 RHEL7 deploys [1], this patch is just making that
recommended work-around the official behavior from now on.

[1] https://access.redhat.com/solutions/5368911

Signed-off-by: Rafael Aquini <aquini@redhat.com>