netfilter: nf_tables: skip deactivated anonymous sets during lookups
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=2196159
Upstream Status: RHEL7 only
CVE: CVE-2023-32233

The fix for the above CVE was incomplete in *RHEL7*.
its not enough to check if the set is scheduled for removal when
an element is supposed to be deleted, this check needs to be done
for all other element operations too, e.g. when an element is
supposed to be *added* to a set.

Move the check to the two functions that do the set lookup.
sets that are scheduled for removal/pending in the transaction
are no longer found if they have the "removed" bit set.

Fixes: ffb7eb4b21c69 ("netfilter: nf_tables: deactivate anonymous set from preparation phase")
Reported-by: Phil Sutter <psutter@redhat.com>
Signed-off-by: Florian Westphal <fwestpha@redhat.com>