selftests: openvswitch: Refactor actions parsing.
Until recently, the ovs-dpctl utility was used with a limited actions set and
didn't need to have support for multiple similar actions.  However, when adding
support for tunnels, it will be important to support multiple set() actions
in a single flow.  When printing these actions, the existing code will be
unable to print all of the sets - it will only print the first.

Refactor this code to be easier to read and support multiple actions of the
same type in an action list.

Signed-off-by: Aaron Conole <aconole@redhat.com>